refactor: Visibilty 이넘 위치 model 하위로 이동
