+++ src/main/java/sws/songpin/domain/playlist/dto/response/PlaylistDetailsResponseDto.java
@@ -1,7 +1,7 @@
 package sws.songpin.domain.playlist.dto.response;
 
 import sws.songpin.domain.genre.entity.GenreName;
-import sws.songpin.domain.pin.entity.Visibility;
+import sws.songpin.domain.model.Visibility;
 import sws.songpin.domain.playlist.entity.Playlist;
 import sws.songpin.domain.song.dto.response.SongInfoDto;
